Menu builder

+++ src/app/Models/PageSectionsPivot.php
@@ -52,6 +52,9 @@ class PageSectionsPivot extends Model
         'data' => 'array',
     ];
 
+    /**
+     * Booted
+     */
     public static function booted()
     {
         // On creating assign a uuid
@@ -131,51 +134,60 @@ class PageSectionsPivot extends Model
      */
     public function setDataAttribute($value)
     {
-        $attribute_name = "data";
+        $image_fields = collect($this->section()->first()->fields)
+            ->filter(function ($item, $key) {
+                return $key === 'image';
+            });
 
-        // or use your own disk, defined in config/filesystems.php
-        $disk = config('backpack.base.root_disk_name');
+        if ($image_fields->count() > 0) {
+            $attribute_name = "data";
 
-        // destination path relative to the disk above
-        $destination_path = "public/uploads";
+            // or use your own disk, defined in config/filesystems.php
+            $disk = config('backpack.base.root_disk_name');
 
-        foreach ($value as $key => $val) {
-            if (is_null($val) || Str::startsWith($val, 'data:image')) {
-                // Decode attribute
-                $attribute = json_decode($this->attributes[$attribute_name], true);
+            // destination path relative to the disk above
+            $destination_path = "public/uploads";
 
-                if ($val == null) {
-                    Storage::disk($disk)->delete('public/' . $attribute[$key]);
+            foreach ($value as $key => $val) {
+                if (is_null($val) || Str::startsWith($val, 'data:image')) {
+                    // Decode attribute
+                    $attribute = json_decode($this->attributes[$attribute_name], true);
 
-                    $attribute[$key] = null;
+                    if ($val == null) {
+                        Storage::disk($disk)->delete('public/' . $attribute[$key]);
 
-                    // set null in the database column
-                    $this->attributes[$attribute_name] = json_encode($attribute);
-                } else if (Str::startsWith($val, 'data:image')) {
-                    // 0. Make the image
-                    $image = Image::make($val)->encode('jpg', 90);
+                        $attribute[$key] = null;
 
-                    // 1. Generate a filename.
-                    $filename = md5($val . time()) . '.jpg';
+                        // set null in the database column
+                        $this->attributes[$attribute_name][$key] = json_encode($attribute);
+                    } else if (Str::startsWith($val, 'data:image')) {
+                        // 0. Make the image
+                        $image = Image::make($val)->encode('jpg', 90);
 
-                    // 2. Store the image on disk.
-                    Storage::disk($disk)->put($destination_path . '/' . $filename, $image->stream());
+                        // 1. Generate a filename.
+                        $filename = md5($val . time()) . '.jpg';
 
-                    // 3. Delete the previous image, if there was one.
-                    Storage::disk($disk)->delete('public/' . $attribute[$key]);
+                        // 2. Store the image on disk.
+                        Storage::disk($disk)->put($destination_path . '/' . $filename, $image->stream());
 
-                    // 4. Save the public path to the database
-                    // but first, remove "public/" from the path, since we're pointing to it
-                    // from the root folder; that way, what gets saved in the db
-                    // is the public URL (everything that comes after the domain name)
-                    $public_destination_path = Str::replaceFirst('public/', '', $destination_path);
+                        // 3. Delete the previous image, if there was one.
+                        Storage::disk($disk)->delete('public/' . $attribute[$key]);
 
-                    $attribute[$key] = $public_destination_path . '/' . $filename;
+                        // 4. Save the public path to the database
+                        // but first, remove "public/" from the path, since we're pointing to it
+                        // from the root folder; that way, what gets saved in the db
+                        // is the public URL (everything that comes after the domain name)
+                        $public_destination_path = Str::replaceFirst('public/', '', $destination_path);
 
-                    $this->attributes[$attribute_name] = json_encode($attribute);
+                        $attribute[$key] = $public_destination_path . '/' . $filename;
+
+                        $this->attributes[$attribute_name][$key] = json_encode($attribute);
+                    }
                 }
-            }
 
+            }
+        } else {
+            $this->attributes['data'] = json_encode($value);
         }
     }
 }
